many tweaks
added camera system v1. technical description below
minor bug fix with destructible objects being destroyed as player died
added respawn delay to avoid stupid deaths because of sudden change
oh yeah and removed those annoying beat ticks because they were redundant and temporary

=============================================================================================================================================

description of camera system
I will check the players position against an outer hitbox, if they cross the hitbox I will assume they are traveling in that direction and move the camera towards a snapping box.
This was mostly explained in the GDC talk. But I also made it so that the cameras movement was proportional to the players offset from the location they first snapped from.
This allows the camera to remain mostly stationary when moving within the hitbox, so it may be hard to tell which directions are snapped to at that moment.

I use a different dimension for vertical and horizontal hitboxes, but the approach is identical.
Also camera lerping. I added that too.

Additionally there is debug code to render the hitbox. As I created the system I find it hard NOT to pay attention to, but you guys should tell me if it feels transparent.

All parameters are in obj_view_manager, and are calculated as proportional to the camera size as set in the room editor. This allows us to zoom out a particular room, and use equally applicable camera controls.